requirements.txt
chardet==5.2.0
pillow==10.2.0
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
python-docx==1.1.0
tqdm==4.66.1
